+++ src/main/java/in/koreatech/koin/domain/shop/repository/MenuRepository.java
@@ -7,15 +7,15 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.shop.exception.MenuNotFoundException;
 import in.koreatech.koin.domain.shop.model.Menu;
 
-public interface MenuRepository extends Repository<Menu, Long> {
+public interface MenuRepository extends Repository<Menu, Integer> {
 
-    Optional<Menu> findById(Long menuId);
+    Optional<Menu> findById(Integer menuId);
 
     Menu save(Menu menu);
 
-    Void deleteById(Long id);
+    Void deleteById(Integer id);
 
-    default Menu getById(Long menuId) {
+    default Menu getById(Integer menuId) {
         return findById(menuId).orElseThrow(() -> MenuNotFoundException.withDetail("menuId: " + menuId));
     }
 }
